+++ src/ethereum_test_forks/forks/pre_merge.py
@@ -1,14 +1,13 @@
 """All Ethereum fork class definitions."""
 
 from dataclasses import replace
-from hashlib import sha256
 from os.path import realpath
 from pathlib import Path
 from typing import List, Mapping, Optional, Sized, Tuple
 
 from semver import Version
 
-from ethereum_test_base_types import AccessList, Address, BlobSchedule, Bytes, ForkBlobSchedule
+from ethereum_test_base_types import AccessList, Address, BlobSchedule, Bytes
 from ethereum_test_base_types.conversions import BytesConvertible
 from ethereum_test_vm import EVMCodeType, Opcodes
 
@@ -22,7 +21,7 @@ from ..base_fork import (
     TransactionIntrinsicCostCalculator,
 )
 from ..gas_costs import GasCosts
-from .helpers import ceiling_division, fake_exponential
+from .helpers import ceiling_division
 
 CURRENT_FILE = Path(realpath(__file__))
 CURRENT_FOLDER = CURRENT_FILE.parent
@@ -802,576 +801,3 @@ class GrayGlacier(ArrowGlacier, solc_name="london", ignore=True):
     """Gray Glacier fork."""
 
     pass
-
-
-class Paris(
-    London,
-    transition_tool_name="Merge",
-    blockchain_test_network_name="Paris",
-):
-    """Paris (Merge) fork."""
-
-    @classmethod
-    def header_prev_randao_required(cls, block_number: int = 0, timestamp: int = 0) -> bool:
-        """Prev Randao is required starting from Paris."""
-        return True
-
-    @classmethod
-    def header_zero_difficulty_required(cls, block_number: int = 0, timestamp: int = 0) -> bool:
-        """Zero difficulty is required starting from Paris."""
-        return True
-
-    @classmethod
-    def get_reward(cls, block_number: int = 0, timestamp: int = 0) -> int:
-        """Paris updates the reward to 0."""
-        return 0
-
-    @classmethod
-    def engine_new_payload_version(
-        cls, block_number: int = 0, timestamp: int = 0
-    ) -> Optional[int]:
-        """From Paris, payloads can be sent through the engine API."""
-        return 1
-
-
-class Shanghai(Paris):
-    """Shanghai fork."""
-
-    @classmethod
-    def header_withdrawals_required(cls, block_number: int = 0, timestamp: int = 0) -> bool:
-        """Withdrawals are required starting from Shanghai."""
-        return True
-
-    @classmethod
-    def engine_new_payload_version(
-        cls, block_number: int = 0, timestamp: int = 0
-    ) -> Optional[int]:
-        """From Shanghai, new payload calls must use version 2."""
-        return 2
-
-    @classmethod
-    def valid_opcodes(
-        cls,
-    ) -> List[Opcodes]:
-        """Return list of Opcodes that are valid to work on this fork."""
-        return [Opcodes.PUSH0] + super(Shanghai, cls).valid_opcodes()
-
-
-class Cancun(Shanghai):
-    """Cancun fork."""
-
-    @classmethod
-    def solc_min_version(cls) -> Version:
-        """Return minimum version of solc that supports this fork."""
-        return Version.parse("0.8.24")
-
-    @classmethod
-    def header_excess_blob_gas_required(cls, block_number: int = 0, timestamp: int = 0) -> bool:
-        """Excess blob gas is required starting from Cancun."""
-        return True
-
-    @classmethod
-    def header_blob_gas_used_required(cls, block_number: int = 0, timestamp: int = 0) -> bool:
-        """Blob gas used is required starting from Cancun."""
-        return True
-
-    @classmethod
-    def header_beacon_root_required(cls, block_number: int = 0, timestamp: int = 0) -> bool:
-        """Parent beacon block root is required starting from Cancun."""
-        return True
-
-    @classmethod
-    def blob_gas_price_calculator(
-        cls, block_number: int = 0, timestamp: int = 0
-    ) -> BlobGasPriceCalculator:
-        """Return a callable that calculates the blob gas price at Cancun."""
-        min_base_fee_per_blob_gas = cls.min_base_fee_per_blob_gas(block_number, timestamp)
-        blob_base_fee_update_fraction = cls.blob_base_fee_update_fraction(block_number, timestamp)
-
-        def fn(*, excess_blob_gas) -> int:
-            return fake_exponential(
-                min_base_fee_per_blob_gas,
-                excess_blob_gas,
-                blob_base_fee_update_fraction,
-            )
-
-        return fn
-
-    @classmethod
-    def excess_blob_gas_calculator(
-        cls, block_number: int = 0, timestamp: int = 0
-    ) -> ExcessBlobGasCalculator:
-        """Return a callable that calculates the excess blob gas for a block at Cancun."""
-        target_blobs_per_block = cls.target_blobs_per_block(block_number, timestamp)
-        blob_gas_per_blob = cls.blob_gas_per_blob(block_number, timestamp)
-        target_blob_gas_per_block = target_blobs_per_block * blob_gas_per_blob
-
-        def fn(
-            *,
-            parent_excess_blob_gas: int | None = None,
-            parent_excess_blobs: int | None = None,
-            parent_blob_gas_used: int | None = None,
-            parent_blob_count: int | None = None,
-        ) -> int:
-            if parent_excess_blob_gas is None:
-                assert parent_excess_blobs is not None, "Parent excess blobs are required"
-                parent_excess_blob_gas = parent_excess_blobs * blob_gas_per_blob
-            if parent_blob_gas_used is None:
-                assert parent_blob_count is not None, "Parent blob count is required"
-                parent_blob_gas_used = parent_blob_count * blob_gas_per_blob
-            if parent_excess_blob_gas + parent_blob_gas_used < target_blob_gas_per_block:
-                return 0
-            else:
-                return parent_excess_blob_gas + parent_blob_gas_used - target_blob_gas_per_block
-
-        return fn
-
-    @classmethod
-    def min_base_fee_per_blob_gas(cls, block_number: int = 0, timestamp: int = 0) -> int:
-        """Return the minimum base fee per blob gas for Cancun."""
-        return 1
-
-    @classmethod
-    def blob_base_fee_update_fraction(cls, block_number: int = 0, timestamp: int = 0) -> int:
-        """Return the blob base fee update fraction for Cancun."""
-        return 3338477
-
-    @classmethod
-    def blob_gas_per_blob(cls, block_number: int = 0, timestamp: int = 0) -> int:
-        """Blobs are enabled starting from Cancun."""
-        return 2**17
-
-    @classmethod
-    def supports_blobs(cls, block_number: int = 0, timestamp: int = 0) -> bool:
-        """At Cancun, blobs support is enabled."""
-        return True
-
-    @classmethod
-    def target_blobs_per_block(cls, block_number: int = 0, timestamp: int = 0) -> int:
-        """Blobs are enabled starting from Cancun, with a static target of 3 blobs."""
-        return 3
-
-    @classmethod
-    def max_blobs_per_block(cls, block_number: int = 0, timestamp: int = 0) -> int:
-        """Blobs are enabled starting from Cancun, with a static max of 6 blobs."""
-        return 6
-
-    @classmethod
-    def blob_schedule(cls, block_number: int = 0, timestamp: int = 0) -> BlobSchedule | None:
-        """
-        At Cancun, the fork object runs this routine to get the updated blob
-        schedule.
-        """
-        parent_fork = cls.parent()
-        assert parent_fork is not None, "Parent fork must be defined"
-        blob_schedule = parent_fork.blob_schedule(block_number, timestamp)
-        if blob_schedule is None:
-            last_blob_schedule = None
-            blob_schedule = BlobSchedule()
-        else:
-            last_blob_schedule = blob_schedule.last()
-        current_blob_schedule = ForkBlobSchedule(
-            target_blobs_per_block=cls.target_blobs_per_block(block_number, timestamp),
-            max_blobs_per_block=cls.max_blobs_per_block(block_number, timestamp),
-            base_fee_update_fraction=cls.blob_base_fee_update_fraction(block_number, timestamp),
-        )
-        if last_blob_schedule is None or last_blob_schedule != current_blob_schedule:
-            blob_schedule.append(fork=cls.__name__, schedule=current_blob_schedule)
-        return blob_schedule
-
-    @classmethod
-    def tx_types(cls, block_number: int = 0, timestamp: int = 0) -> List[int]:
-        """At Cancun, blob type transactions are introduced."""
-        return [3] + super(Cancun, cls).tx_types(block_number, timestamp)
-
-    @classmethod
-    def precompiles(cls, block_number: int = 0, timestamp: int = 0) -> List[Address]:
-        """At Cancun, pre-compile for kzg point evaluation is introduced."""
-        return [Address(0xA)] + super(Cancun, cls).precompiles(block_number, timestamp)
-
-    @classmethod
-    def system_contracts(cls, block_number: int = 0, timestamp: int = 0) -> List[Address]:
-        """Cancun introduces the system contract for EIP-4788."""
-        return [Address(0x000F3DF6D732807EF1319FB7B8BB8522D0BEAC02)]
-
-    @classmethod
-    def pre_allocation_blockchain(cls) -> Mapping:
-        """
-        Cancun requires pre-allocation of the beacon root contract for EIP-4788 on blockchain
-        type tests.
-        """
-        new_allocation = {
-            0x000F3DF6D732807EF1319FB7B8BB8522D0BEAC02: {
-                "nonce": 1,
-                "code": "0x3373fffffffffffffffffffffffffffffffffffffffe14604d57602036146024575f5f"
-                "fd5b5f35801560495762001fff810690815414603c575f5ffd5b62001fff01545f5260205ff35b5f"
-                "5ffd5b62001fff42064281555f359062001fff015500",
-            }
-        }
-        return new_allocation | super(Cancun, cls).pre_allocation_blockchain()  # type: ignore
-
-    @classmethod
-    def engine_new_payload_version(
-        cls, block_number: int = 0, timestamp: int = 0
-    ) -> Optional[int]:
-        """From Cancun, new payload calls must use version 3."""
-        return 3
-
-    @classmethod
-    def engine_new_payload_blob_hashes(cls, block_number: int = 0, timestamp: int = 0) -> bool:
-        """From Cancun, payloads must have blob hashes."""
-        return True
-
-    @classmethod
-    def engine_new_payload_beacon_root(cls, block_number: int = 0, timestamp: int = 0) -> bool:
-        """From Cancun, payloads must have a parent beacon block root."""
-        return True
-
-    @classmethod
-    def valid_opcodes(
-        cls,
-    ) -> List[Opcodes]:
-        """Return list of Opcodes that are valid to work on this fork."""
-        return [
-            Opcodes.BLOBHASH,
-            Opcodes.BLOBBASEFEE,
-            Opcodes.TLOAD,
-            Opcodes.TSTORE,
-            Opcodes.MCOPY,
-        ] + super(Cancun, cls).valid_opcodes()
-
-
-class Prague(Cancun):
-    """Prague fork."""
-
-    @classmethod
-    def is_deployed(cls) -> bool:
-        """
-        Flag that the fork has not been deployed to mainnet; it is under active
-        development.
-        """
-        return False
-
-    @classmethod
-    def solc_min_version(cls) -> Version:
-        """Return minimum version of solc that supports this fork."""
-        return Version.parse("1.0.0")  # set a high version; currently unknown
-
-    @classmethod
-    def precompiles(cls, block_number: int = 0, timestamp: int = 0) -> List[Address]:
-        """
-        At Prague, pre-compile for BLS operations are added.
-
-        G1ADD = 0x0B
-        G1MSM = 0x0C
-        G2ADD = 0x0D
-        G2MSM = 0x0E
-        PAIRING = 0x0F
-        MAP_FP_TO_G1 = 0x10
-        MAP_FP2_TO_G2 = 0x11
-        """
-        return [Address(i) for i in range(0xB, 0x11 + 1)] + super(Prague, cls).precompiles(
-            block_number, timestamp
-        )
-
-    @classmethod
-    def gas_costs(cls, block_number: int = 0, timestamp: int = 0) -> GasCosts:
-        """
-        On Prague, the standard token cost and the floor token costs are introduced due to
-        EIP-7623.
-        """
-        return replace(
-            super(Prague, cls).gas_costs(block_number, timestamp),
-            G_TX_DATA_STANDARD_TOKEN_COST=4,  # https://eips.ethereum.org/EIPS/eip-7623
-            G_TX_DATA_FLOOR_TOKEN_COST=10,
-            G_AUTHORIZATION=25_000,
-            R_AUTHORIZATION_EXISTING_AUTHORITY=12_500,
-        )
-
-    @classmethod
-    def system_contracts(cls, block_number: int = 0, timestamp: int = 0) -> List[Address]:
-        """Prague introduces the system contracts for EIP-6110, EIP-7002, EIP-7251 and EIP-2935."""
-        return [
-            Address(0x00000000219AB540356CBB839CBE05303D7705FA),
-            Address(0x00000961EF480EB55E80D19AD83579A64C007002),
-            Address(0x0000BBDDC7CE488642FB579F8B00F3A590007251),
-            Address(0x0000F90827F1C53A10CB7A02335B175320002935),
-        ] + super(Prague, cls).system_contracts(block_number, timestamp)
-
-    @classmethod
-    def max_request_type(cls, block_number: int = 0, timestamp: int = 0) -> int:
-        """At Prague, three request types are introduced, hence the max request type is 2."""
-        return 2
-
-    @classmethod
-    def calldata_gas_calculator(
-        cls, block_number: int = 0, timestamp: int = 0
-    ) -> CalldataGasCalculator:
-        """
-        Return a callable that calculates the transaction gas cost for its calldata
-        depending on its contents.
-        """
-        gas_costs = cls.gas_costs(block_number, timestamp)
-
-        def fn(*, data: BytesConvertible, floor: bool = False) -> int:
-            tokens = 0
-            for b in Bytes(data):
-                if b == 0:
-                    tokens += 1
-                else:
-                    tokens += 4
-            if floor:
-                return tokens * gas_costs.G_TX_DATA_FLOOR_TOKEN_COST
-            return tokens * gas_costs.G_TX_DATA_STANDARD_TOKEN_COST
-
-        return fn
-
-    @classmethod
-    def transaction_data_floor_cost_calculator(
-        cls, block_number: int = 0, timestamp: int = 0
-    ) -> TransactionDataFloorCostCalculator:
-        """On Prague, due to EIP-7623, the transaction data floor cost is introduced."""
-        calldata_gas_calculator = cls.calldata_gas_calculator(block_number, timestamp)
-        gas_costs = cls.gas_costs(block_number, timestamp)
-
-        def fn(*, data: BytesConvertible) -> int:
-            return calldata_gas_calculator(data=data, floor=True) + gas_costs.G_TRANSACTION
-
-        return fn
-
-    @classmethod
-    def transaction_intrinsic_cost_calculator(
-        cls, block_number: int = 0, timestamp: int = 0
-    ) -> TransactionIntrinsicCostCalculator:
-        """
-        At Prague, the transaction intrinsic cost needs to take the
-        authorizations into account.
-        """
-        super_fn = super(Prague, cls).transaction_intrinsic_cost_calculator(
-            block_number, timestamp
-        )
-        gas_costs = cls.gas_costs(block_number, timestamp)
-        transaction_data_floor_cost_calculator = cls.transaction_data_floor_cost_calculator(
-            block_number, timestamp
-        )
-
-        def fn(
-            *,
-            calldata: BytesConvertible = b"",
-            contract_creation: bool = False,
-            access_list: List[AccessList] | None = None,
-            authorization_list_or_count: Sized | int | None = None,
-            return_cost_deducted_prior_execution: bool = False,
-        ) -> int:
-            intrinsic_cost: int = super_fn(
-                calldata=calldata,
-                contract_creation=contract_creation,
-                access_list=access_list,
-                return_cost_deducted_prior_execution=False,
-            )
-            if authorization_list_or_count is not None:
-                if isinstance(authorization_list_or_count, Sized):
-                    authorization_list_or_count = len(authorization_list_or_count)
-                intrinsic_cost += authorization_list_or_count * gas_costs.G_AUTHORIZATION
-
-            if return_cost_deducted_prior_execution:
-                return intrinsic_cost
-
-            transaction_floor_data_cost = transaction_data_floor_cost_calculator(data=calldata)
-            return max(intrinsic_cost, transaction_floor_data_cost)
-
-        return fn
-
-    @classmethod
-    def blob_base_fee_update_fraction(cls, block_number: int = 0, timestamp: int = 0) -> int:
-        """Return the blob base fee update fraction for Prague."""
-        return 5007716
-
-    @classmethod
-    def target_blobs_per_block(cls, block_number: int = 0, timestamp: int = 0) -> int:
-        """Target blob count of 6 for Prague."""
-        return 6
-
-    @classmethod
-    def max_blobs_per_block(cls, block_number: int = 0, timestamp: int = 0) -> int:
-        """Max blob count of 9 for Prague."""
-        return 9
-
-    @classmethod
-    def pre_allocation_blockchain(cls) -> Mapping:
-        """
-        Prague requires pre-allocation of the beacon chain deposit contract for EIP-6110,
-        the exits contract for EIP-7002, and the history storage contract for EIP-2935.
-        """
-        new_allocation = {}
-
-        # Add the beacon chain deposit contract
-        deposit_contract_tree_depth = 32
-        storage = {}
-        next_hash = sha256(b"\x00" * 64).digest()
-        for i in range(deposit_contract_tree_depth + 2, deposit_contract_tree_depth * 2 + 1):
-            storage[i] = next_hash
-            next_hash = sha256(next_hash + next_hash).digest()
-
-        with open(CURRENT_FOLDER / "contracts" / "deposit_contract.bin", mode="rb") as f:
-            new_allocation.update(
-                {
-                    0x00000000219AB540356CBB839CBE05303D7705FA: {
-                        "nonce": 1,
-                        "code": f.read(),
-                        "storage": storage,
-                    }
-                }
-            )
-
-        # EIP-7002: Add the withdrawal request contract
-        with open(CURRENT_FOLDER / "contracts" / "withdrawal_request.bin", mode="rb") as f:
-            new_allocation.update(
-                {
-                    0x00000961EF480EB55E80D19AD83579A64C007002: {
-                        "nonce": 1,
-                        "code": f.read(),
-                    },
-                }
-            )
-
-        # EIP-7251: Add the consolidation request contract
-        with open(CURRENT_FOLDER / "contracts" / "consolidation_request.bin", mode="rb") as f:
-            new_allocation.update(
-                {
-                    0x0000BBDDC7CE488642FB579F8B00F3A590007251: {
-                        "nonce": 1,
-                        "code": f.read(),
-                    },
-                }
-            )
-
-        # EIP-2935: Add the history storage contract
-        with open(CURRENT_FOLDER / "contracts" / "history_contract.bin", mode="rb") as f:
-            new_allocation.update(
-                {
-                    0x0000F90827F1C53A10CB7A02335B175320002935: {
-                        "nonce": 1,
-                        "code": f.read(),
-                    }
-                }
-            )
-
-        return new_allocation | super(Prague, cls).pre_allocation_blockchain()  # type: ignore
-
-    @classmethod
-    def header_requests_required(cls, block_number: int = 0, timestamp: int = 0) -> bool:
-        """
-        Prague requires that the execution layer header contains the beacon
-        chain requests hash.
-        """
-        return True
-
-    @classmethod
-    def engine_new_payload_requests(cls, block_number: int = 0, timestamp: int = 0) -> bool:
-        """From Prague, new payloads include the requests hash as a parameter."""
-        return True
-
-    @classmethod
-    def engine_new_payload_version(
-        cls, block_number: int = 0, timestamp: int = 0
-    ) -> Optional[int]:
-        """From Prague, new payload calls must use version 4."""
-        return 4
-
-    @classmethod
-    def engine_forkchoice_updated_version(
-        cls, block_number: int = 0, timestamp: int = 0
-    ) -> Optional[int]:
-        """At Prague, version number of NewPayload and ForkchoiceUpdated diverge."""
-        return 3
-
-
-class CancunEIP7692(  # noqa: SC200
-    Cancun,
-    transition_tool_name="Prague",  # Evmone enables (only) EOF at Prague
-    blockchain_test_network_name="Prague",  # Evmone enables (only) EOF at Prague
-    solc_name="cancun",
-):
-    """Cancun + EIP-7692 (EOF) fork (Deprecated)."""
-
-    @classmethod
-    def evm_code_types(cls, block_number: int = 0, timestamp: int = 0) -> List[EVMCodeType]:
-        """EOF V1 is supported starting from this fork."""
-        return super(CancunEIP7692, cls).evm_code_types(  # noqa: SC200
-            block_number,
-            timestamp,
-        ) + [EVMCodeType.EOF_V1]
-
-    @classmethod
-    def call_opcodes(
-        cls, block_number: int = 0, timestamp: int = 0
-    ) -> List[Tuple[Opcodes, EVMCodeType]]:
-        """EOF V1 introduces EXTCALL, EXTSTATICCALL, EXTDELEGATECALL."""
-        return [
-            (Opcodes.EXTCALL, EVMCodeType.EOF_V1),
-            (Opcodes.EXTSTATICCALL, EVMCodeType.EOF_V1),
-            (Opcodes.EXTDELEGATECALL, EVMCodeType.EOF_V1),
-        ] + super(
-            CancunEIP7692,
-            cls,  # noqa: SC200
-        ).call_opcodes(block_number, timestamp)
-
-    @classmethod
-    def create_opcodes(
-        cls, block_number: int = 0, timestamp: int = 0
-    ) -> List[Tuple[Opcodes, EVMCodeType]]:
-        """EOF V1 introduces `EOFCREATE`."""
-        return [(Opcodes.EOFCREATE, EVMCodeType.EOF_V1)] + super(
-            CancunEIP7692,
-            cls,  # noqa: SC200
-        ).create_opcodes(block_number, timestamp)
-
-    @classmethod
-    def is_deployed(cls) -> bool:
-        """
-        Flag that the fork has not been deployed to mainnet; it is under active
-        development.
-        """
-        return False
-
-    @classmethod
-    def solc_min_version(cls) -> Version:
-        """Return minimum version of solc that supports this fork."""
-        return Version.parse("1.0.0")  # set a high version; currently unknown
-
-
-class Osaka(Prague, solc_name="cancun"):
-    """Osaka fork."""
-
-    @classmethod
-    def evm_code_types(cls, block_number: int = 0, timestamp: int = 0) -> List[EVMCodeType]:
-        """EOF V1 is supported starting from Osaka."""
-        return super(Osaka, cls).evm_code_types(
-            block_number,
-            timestamp,
-        ) + [EVMCodeType.EOF_V1]
-
-    @classmethod
-    def call_opcodes(
-        cls, block_number: int = 0, timestamp: int = 0
-    ) -> List[Tuple[Opcodes, EVMCodeType]]:
-        """EOF V1 introduces EXTCALL, EXTSTATICCALL, EXTDELEGATECALL."""
-        return [
-            (Opcodes.EXTCALL, EVMCodeType.EOF_V1),
-            (Opcodes.EXTSTATICCALL, EVMCodeType.EOF_V1),
-            (Opcodes.EXTDELEGATECALL, EVMCodeType.EOF_V1),
-        ] + super(Osaka, cls).call_opcodes(block_number, timestamp)
-
-    @classmethod
-    def is_deployed(cls) -> bool:
-        """
-        Flag that the fork has not been deployed to mainnet; it is under active
-        development.
-        """
-        return False
-
-    @classmethod
-    def solc_min_version(cls) -> Version:
-        """Return minimum version of solc that supports this fork."""
-        return Version.parse("1.0.0")  # set a high version; currently unknown
